feat: refactor GameDetails component and add new custom hook for game management
