+++ src/components/pages/GameDetails/GameDetails.tsx
